requirements.txt
# LangGraph 和 LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# MCP 工具
mcp>=0.1.0

# API 框架
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.9.0

# 数据库
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# 工具库
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.66.0

# 日志和监控
loguru>=0.7.0

# 测试
pytest>=8.0.0
pytest-asyncio>=0.23.0
